% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_question_template.R
\name{add_additional_detail}
\alias{add_additional_detail}
\title{Add Additional Details}
\usage{
add_additional_detail(question_template, detail_name, value)
}
\description{
Add Additional Details
}
