% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_question_template.R
\name{add_answer}
\alias{add_answer}
\title{Add a slot to a question template}
\usage{
add_answer(question_template, text, feedback, mark)
}
\arguments{
\item{question_template}{A question template created with `create_question_template`}

\item{text}{the response}

\item{feedback}{feedback to give if this response was given}

\item{mark}{Value between 0 and 1.}
}
\description{
Add a slot to a question template
}
