% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_question_template.R
\name{add_dataset}
\alias{add_dataset}
\title{Add a dataset resource to a question.}
\usage{
add_dataset(question_template, name, type, value)
}
\arguments{
\item{question_template}{template for a question.}

\item{name}{name the dataset will take within the question.}

\item{type}{type of dataset call (path or function)}

\item{value}{the call.}
}
\description{
Add a dataset resource to a question.
}
