% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_question_template.R
\name{add_preprocessing_script}
\alias{add_preprocessing_script}
\title{Add a preprocessing script to be run for the question.}
\usage{
add_preprocessing_script(question_template, path)
}
\arguments{
\item{question_template}{a template for a question.}

\item{path}{path to script}
}
\description{
Add a preprocessing script to be run for the question.
}
