% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_question_template.R
\name{add_slot}
\alias{add_slot}
\title{Add a slot to a question template}
\usage{
add_slot(question_template, dataset, id, field)
}
\arguments{
\item{question_template}{A question template created with `create_question_template`}

\item{dataset}{The dataset}

\item{id}{The id of the row}

\item{field}{The field of the value}
}
\description{
Add a slot to a question template
}
