% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/af_split.R
\name{af_split}
\alias{af_split}
\title{Internal function used for reading in allele frequency file.}
\usage{
af_split(x, n_vari)
}
\arguments{
\item{x}{character string containing numeric values separated by backslashes.}

\item{n_vari}{number of variants.}
}
\description{
Converts a string to vector of frequencies. e.g. "0.01/0.89" to c(0.01,0.89,0.1)
}
