% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ask_multiple_questions.R
\name{ask_multiple_questions_cmd}
\alias{ask_multiple_questions_cmd}
\title{Ask multiple questions}
\usage{
ask_multiple_questions_cmd(question_templates, n_questions = 1)
}
\arguments{
\item{question_templates}{Set of questions}

\item{n_questions}{Number of questions to ask}
}
\description{
Ask multiple questions
}
