% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ask_question_cmd.R
\name{ask_question_cmd}
\alias{ask_question_cmd}
\title{Ask a question using the command-line.}
\usage{
ask_question_cmd(question)
}
\arguments{
\item{question}{Question.}
}
\description{
Ask a question using the command-line.
}
