% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_.R
\name{create_question}
\alias{create_question}
\title{Create a question from a template.}
\usage{
create_question(template)
}
\arguments{
\item{template}{A question template.}
}
\description{
Create a question from a template.
}
