% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_questions.R
\name{create_question_set}
\alias{create_question_set}
\title{Create questions}
\usage{
create_question_set(question_templates, question_ids)
}
\arguments{
\item{question_templates}{Set of templates}

\item{ids}{character vector of IDs.}
}
\description{
Create questions
}
