% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_question_template.R
\name{create_question_template}
\alias{create_question_template}
\title{Create a template}
\usage{
create_question_template(
  context,
  background,
  prompt,
  preprocessing_script = NULL,
  datasets = NULL
)
}
\arguments{
\item{context}{Context for the conversation}

\item{background}{context for question}

\item{prompt}{instruction to answer the question}

\item{preprocessing_script}{preprocessing script.}

\item{datasets}{The datasets.}
}
\description{
Create a template
}
