% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_questions.R
\name{create_questions}
\alias{create_questions}
\title{Create questions}
\usage{
create_questions(question_templates, n_questions = 10, duplicates = TRUE)
}
\arguments{
\item{question_templates}{Set of templates}

\item{n_questions}{Number of questions to ask}

\item{duplicates}{Allow double-ups?}
}
\description{
Create questions
}
