% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_question_template.R
\name{generate_random_string}
\alias{generate_random_string}
\title{Generate a random string.}
\usage{
generate_random_string(length = 20, character_set = c(LETTERS, letters, 0:9))
}
\arguments{
\item{length}{Number of characters}

\item{character_set}{Vector of characters}
}
\description{
Generate a random string.
}
