% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_.R
\name{get_value}
\alias{get_value}
\title{Internal function for retrieving the value for a slot.}
\usage{
get_value(data, sample_id, slot)
}
\arguments{
\item{data}{data-frame}

\item{sample_id}{Ranodmly selected ID.}

\item{slot}{Slot.}
}
\description{
Internal function for retrieving the value for a slot.
}
