% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret_googlesheets_data.R
\name{interpret_googlesheets_data}
\alias{interpret_googlesheets_data}
\title{Function converting Google Sheet data into questions.}
\usage{
interpret_googlesheets_data(question_data)
}
\arguments{
\item{question_data}{The question data from a row of the google-sheets table.}
}
\description{
Function converting Google Sheet data into questions.
}
