% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_.R
\name{match_answer}
\alias{match_answer}
\title{Determines if a given response matches an answer text.}
\usage{
match_answer(answer_text, response)
}
\arguments{
\item{answer_text}{Text for a given answer.}

\item{response}{user response}
}
\description{
Determines if a given response matches an answer text.
}
