% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_.R
\name{match_templates_by_id}
\alias{match_templates_by_id}
\title{Subset by ID}
\usage{
match_templates_by_id(template_list, ids)
}
\arguments{
\item{template_list}{a list of question templates}

\item{ids}{Ids of the templates to be subsetted.}
}
\description{
Subset by ID
}
