% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_questions_as_md.R
\name{prepare_answer_text}
\alias{prepare_answer_text}
\title{Remove grepl details for answer text.}
\usage{
prepare_answer_text(x)
}
\arguments{
\item{x}{answer text}
}
\description{
Remove grepl details for answer text.
}
