% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_questions_as_md.R
\name{print_answers_as_md}
\alias{print_answers_as_md}
\title{Print answers to question}
\usage{
print_answers_as_md(question)
}
\arguments{
\item{question}{A question}
}
\description{
Print answers to question
}
