% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_questions_as_md.R
\name{print_questions_as_md}
\alias{print_questions_as_md}
\title{Print questions as a list of markdown questions}
\usage{
print_questions_as_md(questions)
}
\arguments{
\item{questions}{Question set}
}
\description{
Print questions as a list of markdown questions
}
