% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret_googlesheets_data.R
\name{process_ids_from_string}
\alias{process_ids_from_string}
\title{Internal function for processing IDs.}
\usage{
process_ids_from_string(x)
}
\arguments{
\item{x}{Ids as string.}
}
\description{
Internal function for processing IDs.
}
