% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_questions_as_html.R
\name{render_questions_as_html}
\alias{render_questions_as_html}
\title{Render questions as some HTML}
\usage{
render_questions_as_html(questions, output = tempfile(fileext = ".html"))
}
\arguments{
\item{questions}{Questions to be rendered.}

\item{output}{file path to output result to.}
}
\description{
Render questions as some HTML
}
