% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ask_multiple_questions.R
\name{summarise_question_results}
\alias{summarise_question_results}
\title{Summarise the results of `ask_multiple_questions_cmd`}
\usage{
summarise_question_results(answers)
}
\arguments{
\item{answers}{Results}
}
\description{
Summarise the results of `ask_multiple_questions_cmd`
}
