% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_question_template.R
\name{get_additional_details}
\alias{get_additional_details}
\title{Get Additional Details}
\usage{
get_additional_details(question_templates, detail_name)
}
\arguments{
\item{question_templates}{the question template}

\item{detail_name}{the detail name}
}
\description{
Get Additional Details
}
